# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stadion Tracker API",
        "description": "API-documentatie voor stadionbezoeken, wedstrijden, posts en gebruikersinteractie.",
        "contact": {
            "email": "cedricvanhoorebeke@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.ddev.site",
            "description": "DDEV ontwikkelomgeving"
        }
    ],
    "paths": {
...
        "/api/likes": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Voeg een like toe aan een post",
                "operationId": "511c8161eb44f7b3479a914692d975ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post geliket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Niet geauthenticeerd"
                    },
                    "422": {
                        "description": "Validatiefout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/likes/{post_id}": {
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Verwijder een like van een post",
                "operationId": "93f47e540787d577a30a2188b6fcfb66",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "ID van de post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like verwijderd",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Like niet gevonden"
                    },
                    "401": {
                        "description": "Niet geauthenticeerd"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Comment": {
                "required": [
                    "user_id",
                    "post_id",
                    "comment"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 12
                    },
                    "comment": {
                        "type": "string",
                        "example": "Leuke sfeer tijdens de match!"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:05:00Z"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Game": {
                "required": [
                    "stadium_id",
                    "home_team_id",
                    "away_team_id",
                    "match_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "stadium_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "home_team_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "away_team_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "match_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-14"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:05:00Z"
                    },
                    "stadium": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Stadium"
                            }
                        ],
                        "nullable": true
                    },
                    "home_team": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Team"
                            }
                        ],
                        "nullable": true
                    },
                    "away_team": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Team"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Like": {
                "required": [
                    "user_id",
                    "post_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 10
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-14T10:15:00Z"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    },
                    "post": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Post"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "required": [
                    "user_id",
                    "game"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "game": {
                        "$ref": "#/components/schemas/Game"
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    },
                    "likes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Like"
                        }
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/posts/12345.jpg",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "example": "FC Barcelona vs Real Madrid",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T14:20:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T15:00:00Z"
                    }
                },
                "type": "object"
            },
            "Stadium": {
                "required": [
                    "name",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "name": {
                        "type": "string",
                        "example": "Allianz Arena"
                    },
                    "team_name": {
                        "type": "string",
                        "example": "Bayern Munich",
                        "nullable": true
                    },
                    "profile_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/uploads/stadiums/profile-image/abc.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/uploads/stadiums/banner-image/xyz.jpg",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "string",
                        "example": "48.2188"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "11.6247"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Team": {
                "required": [
                    "name",
                    "league_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Manchester City"
                    },
                    "league_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "profile_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/teams/profile-image/abc.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/teams/banner-image/xyz.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:30:00Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "id",
                    "name",
                    "username",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cédric Van Hoorebeke"
                    },
                    "username": {
                        "type": "string",
                        "example": "cedrictest"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "cedric@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user"
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "https://s3-url.com/image.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "example": "https://s3-url.com/banner.jpg",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-18T12:34:56Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-18T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-18T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}